refactor: extract product operations into custom hook and rename loading/error states
